backend/requirements.txt
fastapi            #Backend framework
uvicorn            #Runs the server
openai             #AI itinerary generation
python-dotenv      #Loads API key from .env